+++ gskewed/support Classes/Operators.cs
@@ -8,12 +8,12 @@ namespace gskewed.support_Classes
 {
     class Operators
     {
-        public static int XOR(int a, int b){
+        public static long XOR(long a, long b){
             return a ^ b;
         }
 
-        public static int concat(int a, int b) {
-            return Convert.ToInt32(Convert.ToString(a,2) + Convert.ToString(b,2),2);
+        public static long concat(long a, long b) {
+            return Convert.ToInt64(Convert.ToString(a,2) + Convert.ToString(b,2),2);
         }
 
         public static bool NOR(bool a, bool b)
